refactor reorder panel strategy location for fewer files
