+++ app/strategies/component/dashboard-panel/update-entity-order.strategy.ts
